fix: separete errorModal
